src/requirements.txt
grpcio
grpcio-tools
redis
dnspython
